requirements.txt
numpy==1.23.5
pandas==1.5.3
pandas-datareader==0.10.0
pandas-gbq==0.17.9
pandas-stubs==1.5.3.230304
pandocfilters==1.5.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.11.3
tensorboard==2.13.0
tensorflow==2.13.0rc1
transformers==4.35.2
torch==2.1.0
torchaudio==2.1.0
torchdata==0.7.0
torchsummary==1.5.1
torchtext==0.16.0
torchvision==0.16.0
text-unidecode==1.3
spanish_nlp==0.2.11
accelerate==0.20.3
pyreadr==0.5.0